Function to server-render bento components
